# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: Sick-Fansubs HTTP API
  version: "1"
  description: >-
    Accepted Sick-Fansubs wire contracts. Product APIs use /api/v1; operational
    health endpoints are intentionally outside product API versioning.

    Cross-cutting: every endpoint whose chain carries the forced-change gate
    (the authenticated reads chain and the authenticated chain) additionally
    answers 403 with problem type /problems/auth/password-change-required for
    sessions whose account carries the forced-change flag. The /api/v1/auth/
    subtree never carries the gate, so its paths stay reachable while flagged —
    including PUT /api/v1/auth/password, the sign-out pair, the session
    bootstrap, POST /api/v1/auth/reset-password, and
    POST /api/v1/auth/resend-verification.

    Cross-cutting: every documented unsafe operation (POST, PUT, PATCH,
    DELETE) requires the trusted origin and, when a session is present, the
    session-bound X-CSRF-Token header matching the csrfToken returned by the
    session bootstrap — a missing, malformed, or mismatched token is a 403.
    Safe methods never require it, and an unauthenticated unsafe request
    (for example sign-in, register, forgot-password, or reset-password) is
    protected by the trusted origin alone. An expired session has the
    clearing sf_session cookie written before its handler runs, and the
    operations whose handler clears the cookie — sign-out, sign-out-all,
...
    their own response.

...
servers:
...
paths:
...
  /api/v1/push-subscriptions:
    get:
      operationId: listPushSubscriptions
      summary: List own push subscriptions (authenticated)
      description: >-
        The user's subscription rows (id, endpoint, createdAt) — the
        settings list. Own rows only; oldest first. Floor: authenticated
        (401 unauthenticated). No query parameters are accepted.
      responses:
        "200":
          description: The user's subscription rows (may be empty).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscriptionList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createPushSubscription
      summary: Store or refresh one device subscription (authenticated)
      description: >-
        Stores the browser's PushSubscription JSON (endpoint + p256dh/auth
        keys). 201 + the stored row id — the id is the existing row's id
        when the same endpoint resubscribes (the key material and
        updatedAt refresh in place). An endpoint owned by ANOTHER user is
        reassigned (the device's newest login owns it). The endpoint must
        be an absolute https URL; p256dh/auth must be URL-safe base64
        decoding to exactly 65 and 16 bytes (the RFC 8291 shapes) —
        violations are 422 {field, code}. Unknown body fields are 400;
        body hard-limit 4 KiB (413 beyond). Requires the trusted origin and
        a valid X-CSRF-Token (the authenticated chain).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionCreate"
      responses:
        "201":
          description: The subscription row is stored.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscriptionCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/Validation"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/push-subscriptions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: >-
          Opaque subscription row id (the create response's id). A
          structurally invalid id is a 400.
        schema:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Za-z0-9._~-]{1,64}$"
    delete:
      operationId: deletePushSubscription
      summary: Remove one subscription row (authenticated)
      description: >-
        204 always — the operation is idempotent and user-scoped, and
        unknown/foreign ids are deliberately MASKED (no existence
        disclosure). Requires the trusted origin and a valid X-CSRF-Token
        (the authenticated chain).
      responses:
        "204":
          description: The subscription row (if any) is removed.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/notification-preferences/{kind}:
    parameters:
      - name: kind
        in: path
        required: true
        description: >-
          The notification kind (heart | comment_reply | comment |
          content_updated | new_content | comment_removed |
          draft_activity). An unknown kind is the masked 404 — a path
          resource — and draft_activity answers the SAME 404 to a session
          below the moderator floor.
        schema:
          type: string
          enum:
            - heart
            - comment_reply
            - comment
            - content_updated
            - new_content
            - comment_removed
            - draft_activity
    put:
      operationId: setNotificationPreference
      summary: Write one explicit push toggle (authenticated)
      description: >-
        Stores an explicit per-kind push toggle (204) — a stored row
        overrides the role default from then on. Repeats overwrite
        (last-write-wins; boolean toggles carry no ETag). Below the
        moderator floor the staff-only draft_activity answers the masked 404
        — the same 404 an unknown kind gets. Requires the trusted origin
        and a valid X-CSRF-Token (the authenticated chain). Body hard-limit
        4 KiB (413 beyond); a missing pushEnabled is 422 {pushEnabled,
        required}.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferenceUpdate"
      responses:
        "204":
          description: The toggle is stored.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/Validation"
        "500":
          $ref: "#/components/responses/InternalError"
  # ── Comments ──
...
components:
  headers:
    RequestID:
      description: Server-generated request correlation identifier.
      schema:
        type: string
    NoStoreCacheControl:
      description: API responses default to no-store.
      schema:
        type: string
        const: no-store
...
  schemas:
...
    Problem:
      type: object
      description: >-
        RFC 9457 Problem Details with only accepted safe extensions. Endpoint
        contracts define their stable problem types and violation codes.
      additionalProperties: false
      required:
        - type
        - title
        - status
        - requestId
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable relative problem type identifier.
          example: /problems/validation
        title:
          type: string
          description: Safe developer-oriented summary; not localized UI copy.
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: HTTP status code for this occurrence.
        requestId:
          type: string
          description: Server-generated request correlation identifier.
        violations:
          type: array
          description: Deterministically ordered field violations when applicable.
          items:
            type: object
            additionalProperties: false
            required:
              - field
              - code
            properties:
              field:
                type: string
                description: Accepted public request-field identifier.
              code:
                type: string
                description: Stable machine-readable violation identifier.
...
    PushSubscriptionCreate:
      type: object
      description: >-
        The browser's PushSubscription JSON body: the push-service endpoint
        and the RFC 8291 key material. p256dh/auth must be URL-safe base64
        decoding to exactly 65 and 16
        bytes.
      additionalProperties: false
      required:
        - endpoint
        - keys
      properties:
        endpoint:
          type: string
          minLength: 1
          maxLength: 512
          description: Absolute https URL of the push service endpoint.
        keys:
          type: object
          additionalProperties: false
          required:
            - p256dh
            - auth
          properties:
            p256dh:
              type: string
              maxLength: 512
              description: URL-safe base64 decoding to the 65-byte uncompressed P-256 point.
            auth:
              type: string
              maxLength: 512
              description: URL-safe base64 decoding to the 16-byte authentication secret.
    PushSubscriptionCreated:
      type: object
      description: The create response — the stored row id.
      additionalProperties: false
      required:
        - id
      properties:
        id:
          type: string
          description: >-
            Opaque subscription row id (the existing row's id on a
            same-endpoint resubscribe).
    PushSubscriptionList:
      type: object
      description: The settings-list envelope — own subscription rows only.
      additionalProperties: false
      required:
        - subscriptions
      properties:
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/PushSubscriptionItem"
    PushSubscriptionItem:
      type: object
      description: One subscription row for the settings list.
      additionalProperties: false
      required:
        - id
        - endpoint
        - createdAt
      properties:
        id:
          type: string
        endpoint:
          type: string
          description: >-
            The push-service endpoint. The list carries it so the client
            can match rows against the browser's local registration.
        createdAt:
          type: string
          format: date-time
          description: Subscription instant, canonical UTC RFC 3339 with milliseconds.
...
    NotificationPreferenceUpdate:
      type: object
      description: The PUT body — one explicit toggle.
      additionalProperties: false
      required:
        - pushEnabled
      properties:
        pushEnabled:
          type: boolean
...
  responses:
    BadRequest:
      description: Malformed request (400) — problem type /problems/bad-request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InvalidCredentials:
      description: Generic 401 — problem type /problems/auth/invalid-credentials.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Forbidden:
      description: >-
        403 — problem type /problems/forbidden (CSRF/origin failure or a role
        floor) or /problems/auth/password-change-required (the forced-change
        gate — see PasswordChangeRequired).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    PayloadTooLarge:
      description: 413 body over the endpoint's documented body limit — problem type /problems/payload-too-large.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    UnsupportedMediaType:
      description: 415 non-JSON Content-Type — problem type /problems/unsupported-media-type.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: >-
        404 — problem type /problems/not-found: an unknown path under /api/v1
        (namespace catch-all), or an absent target on a resource operation
        (the operation's description states what its 404 discloses).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Validation:
      description: 422 with field violations — problem type /problems/validation.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    InternalError:
      description: 500 generic unexpected failure — problem type /problems/internal-error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"